requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1

langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-groq==0.1.4

sentence-transformers==2.3.1
faiss-cpu==1.13.2

huggingface-hub==0.21.4

streamlit==1.31.0
plotly==5.18.0
requests==2.31.0
pandas==2.2.0

python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
numpy==1.26.4
groq==0.18.0
PyPDF2==3.0.1
reportlab==4.0.9